Change the rotation
It was originally 180 because I thought it was degrees. It is now changed to 55

The length of the robot is 18 inches, so the back wheel is the center point in the circle and that means that the radius must be 18 inches.
Circumference is 2 pi r. But since it's only 180 degrees, you get pi r.
18 x 3.14 is 56.52, which is rounded to 55.

That is my math, if it is actually measured in degrees and there was actually just an error while running Autonomous, then change it back.