Rewrite some ResQRobot javadoc with new style
@Thunderbots5604/contributors

I put all the declarations on one line because the `private static final double` will propogate to all of the variables in the statement. The javadoc will also propogate to all the variables, which means that we only have to write the javadoc once, and it will be applied to all the variables. Before this change, the javadoc was written once, and only applied to the first of the 4 variables.