Test if the order of registration matters
After attempting all other methods we could think of yesterday, we would now like to try to isolate whether the problem is with First code, or with our code. It is possible that the error is within First's method of registering servos, seeing as changing the servos to ones that were known to work had no effect on the problem. So far we have determined that the servos are going to the correct positions according to the First Servo driver (within the accuracy of 0.1), but are not moving in the physical world. We have absolutely no idea what is going wrong with the system right now, and there appears to have been no record of this problem before.